test: Move job bundle tests to automated tests.
