--- job_bundle_output_tests/physical/expected_job_bundle/asset_references.yaml
@@ -1,9 +1,0 @@
-assetReferences:
-  inputs:
-    directories: []
-    filenames:
-    - C:\normalized\job\bundle\dir\physical.c4d
-  outputs:
-    directories:
-    - C:\normalized\job\bundle\dir\renders
-  referencedPaths: []
