--- job_bundle_output_tests/physical_textured/expected_job_bundle/asset_references.yaml
@@ -1,10 +1,0 @@
-assetReferences:
-  inputs:
-    directories: []
-    filenames:
-    - C:\normalized\job\bundle\dir\physical_textured.c4d
-    - C:\normalized\job\bundle\dir\tex\checkerboard.bmp
-  outputs:
-    directories:
-    - C:\normalized\job\bundle\dir\renders
-  referencedPaths: []
