--- job_bundle_output_tests/physical_textured/expected_job_bundle/parameter_values.yaml
@@ -1,17 +1,0 @@
-parameterValues:
-- name: Cinema4DFile
-  value: C:\normalized\job\bundle\dir\physical_textured.c4d
-- name: OutputPath
-  value: C:\normalized\job\bundle\dir\renders\physical_textured
-- name: MultiPassPath
-  value: ''
-- name: Frames
-  value: '1'
-- name: deadline:targetTaskRunStatus
-  value: READY
-- name: deadline:maxFailedTasksCount
-  value: 20
-- name: deadline:maxRetriesPerTask
-  value: 5
-- name: deadline:priority
-  value: 50
