--- job_bundle_output_tests/redshift_takes/expected_job_bundle/parameter_values.yaml
@@ -1,17 +1,0 @@
-parameterValues:
-- name: Cinema4DFile
-  value: C:\normalized\job\bundle\dir\redshift_takes.c4d
-- name: OutputPath
-  value: C:\normalized\job\bundle\dir\renders\redshift_takes_$take
-- name: MultiPassPath
-  value: ''
-- name: Frames
-  value: '1'
-- name: deadline:targetTaskRunStatus
-  value: READY
-- name: deadline:maxFailedTasksCount
-  value: 20
-- name: deadline:maxRetriesPerTask
-  value: 5
-- name: deadline:priority
-  value: 50
