+++ test/integ/test_scenes/redshift_takes/expected_job_bundle/parameter_values.yaml
@@ -1,0 +1,21 @@
+parameterValues:
+- name: Cinema4DFile
+  value: PATH_TO_BE_REPLACED\deadline-cloud-for-cinema-4d\test\integ\test_scenes\redshift_takes\generated_bundle\redshift_takes.c4d
+- name: OutputPath
+  value: PATH_TO_BE_REPLACED\deadline-cloud-for-cinema-4d\test\integ\test_scenes\redshift_takes\generated_bundle\renders\redshift_takes_$take
+- name: MultiPassPath
+  value: ''
+- name: Frames
+  value: '1'
+- name: CondaPackages
+  value: cinema4d=2025.* cinema4d-openjd=0.6.*
+- name: CondaChannels
+  value: deadline-cloud
+- name: deadline:targetTaskRunStatus
+  value: READY
+- name: deadline:maxFailedTasksCount
+  value: 20
+- name: deadline:maxRetriesPerTask
+  value: 5
+- name: deadline:priority
+  value: 50
