+++ test/integ/test_scenes/redshift_textured/expected_job_bundle/asset_references.yaml
@@ -1,0 +1,10 @@
+assetReferences:
+  inputs:
+    directories: []
+    filenames:
+    - PATH_TO_BE_REPLACED\deadline-cloud-for-cinema-4d\test\integ\test_scenes\redshift_textured\generated_bundle\redshift_textured.c4d
+    - PATH_TO_BE_REPLACED\deadline-cloud-for-cinema-4d\test\integ\test_scenes\redshift_textured\generated_bundle\tex\checkerboard.bmp
+  outputs:
+    directories:
+    - PATH_TO_BE_REPLACED\deadline-cloud-for-cinema-4d\test\integ\test_scenes\redshift_textured\generated_bundle\renders
+  referencedPaths: []
