+++ test/integ/test_scenes/redshift_textured_with_nonascii_characters/expected_job_bundle/parameter_values.yaml
@@ -8,7 +8,7 @@ parameterValues:
     test_scenes\\redshift_textured_with_nonascii_characters\\generated_bundle\\renders\\\
     redshift_textured-_\u20BF_\u0119_\xF1_\u03B2_\u0411_\u062A"
 - name: MultiPassPath
-  value: PATH_TO_BE_REPLACED\deadline-cloud-for-cinema-4d\test\integ\test_scenes\redshift_textured_with_nonascii_characters\generated_bundle
+  value: ''
 - name: Frames
   value: '1'
 - name: CondaPackages
